refactor(logger): unify log and error-path emission into a single emit()
log() and handleHttpError() each reimplemented the same filter/context/redact/transports/file/console pipeline by copy, and had drifted: the error path called logToTransports unconditionally, ignored isEffectivelyDisabled, hardcoded console.error, and never threaded precomputed duration/URL parts. Both now call the same logger/emit.ts pipeline gated by one resolveSinks() result, so a 4xx WARNING now logs via console.warn like the success path, and the error path reuses the single-parse optimization from plan 012.
